refactor(frontend): consolidate stock adjustment hooks into useStockAdjustmentsWorkspace
